Bug resolve and some improovments
In case of context without observation, if new context are loaded, map is not reset. Resolved
MainControlComponent cannot exit from map rectangular area. The draggable plugin is not working as clean as I want, TODO: implement our own plugin
Augmented opacity of MainControl observation area and context
Update Stomp library and resolved a warning with protocols (in a horrible way, but it works, sorry)